requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
attrs==26.1.0
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fastf1==3.8.2
fonttools==4.62.1
greenlet==3.3.2
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
msgpack==1.1.2
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
RapidFuzz==3.14.5
requests==2.33.1
requests-cache==1.3.1
scipy==1.17.1
signalrcore==1.0.2
six==1.17.0
SQLAlchemy==2.0.49
stack-data==0.6.3
timple==0.1.8
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
url-normalize==2.2.1
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
